feat: changes suggested by @magicmatatjahu

+++ src/constants.ts
@@ -40,11 +40,11 @@ export class ContextTestingHelper {
 		return this._context.store[key];
 	}
 
-	createSpecFileAtWorkingDir(){
-		fs.writeFileSync(path.resolve(process.cwd(), 'asyncapi.yml'), '');
+	createSpecFileAtWorkingDir() {
+		if (!fs.existsSync(path.resolve(process.cwd(), 'asyncapi.yml'))) fs.writeFileSync(path.resolve(process.cwd(), 'asyncapi.yml'), '');
 	}
 
-	deleteSpecFileAtWorkingDir(){
-		if(fs.existsSync(path.resolve(process.cwd(), 'asyncapi.yml'))) fs.unlinkSync(path.resolve(process.cwd(), 'asyncapi.yml'));
+	deleteSpecFileAtWorkingDir() {
+		if (fs.existsSync(path.resolve(process.cwd(), 'asyncapi.yml'))) fs.unlinkSync(path.resolve(process.cwd(), 'asyncapi.yml'));
 	}
 }
